refactor(oracles/borrow-rates): Move `MyProvider` and rpc urls to `types`
