+++ apps/oracles/borrow-rates/src/providers/mod.rs
@@ -3,3 +3,4 @@ pub mod hyperdrive;
 pub mod hyperlend;
 pub mod hypurrfi;
 pub mod pool_data_provider;
+pub mod types;
