+++ apps/oracles/borrow-rates/src/providers/pool_data_provider.rs
@@ -7,10 +7,11 @@ use anyhow::Result;
 use blocksense_sdk::eth_rpc::eth_call;
 use url::Url;
 
-use crate::{domain::BorrowRateInfo, utils::math::ray_to_apr};
-
-const RPC_URL: &str = "https://rpc.hyperliquid.xyz/evm";
-pub const ETHEREUM_MAINNET_RPC_URL: &str = "https://eth.llamarpc.com";
+use crate::{
+    domain::BorrowRateInfo,
+    providers::types::{MyProvider, RPC_URL_HYPERLIQUID_MAINNET},
+    utils::math::ray_to_apr,
+};
 
 #[derive(Debug, Clone)]
 pub struct ReserveLike {
@@ -35,7 +36,7 @@ pub trait UiPool {
 
 /// One generic constructor used everywhere.
 pub fn plan_for<P: UiPool>(ui: Address, addresses_provider: Address) -> Result<Plan> {
-    let rpc_url = Url::parse(RPC_URL)?;
+    let rpc_url = Url::parse(RPC_URL_HYPERLIQUID_MAINNET)?;
     let provider = ProviderBuilder::new().connect_http(rpc_url.clone());
     let calldata = P::calldata(provider, ui, addresses_provider);
     Ok(Plan {
@@ -46,9 +47,13 @@ pub fn plan_for<P: UiPool>(ui: Address, addresses_provider: Address) -> Result<P
 }
 
 pub async fn fetch_reserves(plan: Plan) -> Result<Vec<BorrowRateInfo>> {
-    let raw = eth_call(RPC_URL, &format!("{:?}", plan.to), &plan.calldata)
-        .await
-        .map_err(|e| anyhow::Error::msg(format!("{:?}", e)))?;
+    let raw = eth_call(
+        RPC_URL_HYPERLIQUID_MAINNET,
+        &format!("{:?}", plan.to),
+        &plan.calldata,
+    )
+    .await
+    .map_err(|e| anyhow::Error::msg(format!("{:?}", e)))?;
     let reserves = (plan.decode)(&raw)?;
     let mut out = Vec::with_capacity(reserves.len());
     for r in reserves {
@@ -61,20 +66,3 @@ pub async fn fetch_reserves(plan: Plan) -> Result<Vec<BorrowRateInfo>> {
     }
     Ok(out)
 }
-
-pub type MyProvider = alloy::providers::fillers::FillProvider<
-    alloy::providers::fillers::JoinFill<
-        alloy::providers::Identity,
-        alloy::providers::fillers::JoinFill<
-            alloy::providers::fillers::GasFiller,
-            alloy::providers::fillers::JoinFill<
-                alloy::providers::fillers::BlobGasFiller,
-                alloy::providers::fillers::JoinFill<
-                    alloy::providers::fillers::NonceFiller,
-                    alloy::providers::fillers::ChainIdFiller,
-                >,
-            >,
-        >,
-    >,
-    alloy::providers::RootProvider,
->;
