+++ apps/oracles/borrow-rates/src/providers/types.rs
@@ -1,0 +1,19 @@
+pub const RPC_URL_HYPERLIQUID_MAINNET: &str = "https://rpc.hyperliquid.xyz/evm";
+pub const RPC_URL_ETHEREUM_MAINNET: &str = "https://eth.llamarpc.com";
+
+pub type MyProvider = alloy::providers::fillers::FillProvider<
+    alloy::providers::fillers::JoinFill<
+        alloy::providers::Identity,
+        alloy::providers::fillers::JoinFill<
+            alloy::providers::fillers::GasFiller,
+            alloy::providers::fillers::JoinFill<
+                alloy::providers::fillers::BlobGasFiller,
+                alloy::providers::fillers::JoinFill<
+                    alloy::providers::fillers::NonceFiller,
+                    alloy::providers::fillers::ChainIdFiller,
+                >,
+            >,
+        >,
+    >,
+    alloy::providers::RootProvider,
+>;
